requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
scikit-learn==1.3.0

# Financial Data
yfinance==0.2.28
pandas-datareader==0.10.0
sec-edgar-downloader==5.0.3

# NLP & ML (Optional - install for advanced features)
# transformers==4.30.2
# torch==2.0.1
# For CPU-only PyTorch, use: torch==2.0.1+cpu

# Web Scraping & Parsing
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Social Media (Optional - requires API keys)
# tweepy==4.14.0
# praw==7.7.0

# Statistics
statsmodels==0.14.0

# Utilities
pyyaml==6.0
tqdm==4.65.0
python-dotenv==1.0.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Jupyter (Optional - for notebooks)
# jupyter==1.0.0
# ipykernel==6.25.0
# notebook==7.0.0
